refactor: consolidate authorization module into deref step in crd module and build step in controller/build
