--- rust/operator-binary/src/authorization/mod.rs
@@ -1,1 +1,0 @@
-pub mod opa;
